Fixed portfolio list.
Two potential problems:
1. If the use does not have any portfolio, there will be an error.
2. The user needs to be able to switch from one portfolio to the other.
I think we might want to store which portfolio the user visited in the
database (Just an index or the name of it). So when the user first
entered, read that information and display the user's most recent
portfolio.When the user switch from one portfolio to another, update
that.